requirements.txt
Flask
Flask-Login
Flask-Mail
Flask-WTF
mysql-connector-python
Pillow
pygame